Update referenced URL in the gemspec description
The existing linked URL had a bunch of email newsletter stuff in it, so I went and got the 'true' URL of the article instead. In the future, when the article breaks, it may also help in googling for its new location.

Also added the summary into the description, since the description is the only thing that seems to appear on the rubygems web UI. The summary looks like it may be only for gem list -d.